+++ tests/test_live_client_reconnect.py
@@ -65,9 +65,9 @@ async def test_reconnect_before_start(
     reconnect_policy: ReconnectPolicy = ReconnectPolicy.RECONNECT,
 ) -> None:
     """
-    Test that a reconnect policy of "reconnect_do_not_replay" reconnects a
-    client but does not send the session start command if the session was not
-    streaming previously.
+    Test that a reconnect policy of "reconnect" reconnects a client but does
+    not send the session start command if the session was not streaming
+    previously.
     """
     # Arrange
     live_client = client.Live(
@@ -138,8 +138,8 @@ async def test_reconnect_subscriptions(
     reconnect_policy: ReconnectPolicy = ReconnectPolicy.RECONNECT,
 ) -> None:
     """
-    Test that a reconnect policy of "reconnect_do_not_replay" re-sends the
-    subscription requests with a start of `None`.
+    Test that a reconnect policy of "reconnect" re-sends the subscription
+    requests with a start of `None`.
     """
     # Arrange
     live_client = client.Live(
@@ -192,8 +192,8 @@ async def test_reconnect_callback(
     reconnect_policy: ReconnectPolicy = ReconnectPolicy.RECONNECT,
 ) -> None:
     """
-    Test that a reconnect policy of "reconnect_do_not_replay" will cause a user
-    supplied reconnection callback to be executed when a reconnection occurs.
+    Test that a reconnect policy of "reconnect" will cause a user supplied
+    reconnection callback to be executed when a reconnection occurs.
     """
     # Arrange
     live_client = client.Live(